+++ domainservice/src/test/kotlin/systems/ajax/malov/stockanalyzer/mongock/changelogs/UserTrackedSymbolCollectionIndexesMigrationTest.kt
@@ -8,7 +8,6 @@ import systems.ajax.malov.stockanalyzer.util.IntegrationTestBase
 import kotlin.test.Test
 import kotlin.test.assertNull
 
-
 class UserTrackedSymbolCollectionIndexesMigrationTest : IntegrationTestBase() {
 
     @Autowired
