+++ domainservice/src/test/kotlin/systems/ajax/malov/stockanalyzer/util/IntegrationTestBase.kt
@@ -1,8 +1,34 @@
 package systems.ajax.malov.stockanalyzer.util
 
+import org.apache.kafka.clients.admin.Admin
+import org.apache.kafka.clients.admin.KafkaAdminClient
+import org.apache.kafka.common.serialization.ByteArrayDeserializer
+import org.apache.kafka.common.serialization.StringDeserializer
+import org.springframework.boot.autoconfigure.kafka.KafkaProperties
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Import
+import org.springframework.kafka.core.ConsumerFactory
+import org.springframework.kafka.core.DefaultKafkaConsumerFactory
+import org.springframework.kafka.core.KafkaAdmin
 import org.springframework.test.context.ActiveProfiles
 
 @SpringBootTest
 @ActiveProfiles("test")
-abstract class IntegrationTestBase
+@SuppressWarnings("UnnecessaryAbstractClass")
+@Import(IntegrationTestBase.MyKafkaTestConfiguration::class)
+abstract class IntegrationTestBase {
+    class MyKafkaTestConfiguration {
+        @Bean
+        fun adminClient(kafkaAdmin: KafkaAdmin): Admin =
+            KafkaAdminClient.create(kafkaAdmin.configurationProperties)
+
+        @Bean
+        fun consumerFactory(kafkaProperties: KafkaProperties): ConsumerFactory<String, ByteArray> =
+            DefaultKafkaConsumerFactory(
+                kafkaProperties.buildConsumerProperties(null),
+                StringDeserializer(),
+                ByteArrayDeserializer()
+            )
+    }
+}
